Created a ProjectDataHub class for the GUI. Inserted it into our current GUI in Main.java. Also made minor edits to parameters of IssueTable constructor.

+++ application/IssueTable.java
@@ -112,14 +112,15 @@ public class IssueTable extends ScrollPane {
 		}
 	}
 	
-	IssueTable(int prefWidth, int prefHeight, ArrayList<Issue> issues){
+	IssueTable(double prefWidth, double prefHeight, ArrayList<Issue> issues){
 		
 		table = new TableView();
 		
 		table.setEditable(false);
-		table.setPrefHeight((double) prefHeight);
-		table.setPrefWidth((double) prefWidth);
-		
+		table.setPrefHeight(prefHeight);
+		table.setPrefWidth(prefWidth);
+		//this.setPrefHeight(prefHeight);
+		//this.setPrefWidth(prefWidth);
 		TableColumn<Rectangle, Issue> priorityColumn = new TableColumn<>("Priority");
 		priorityColumn.setCellValueFactory(new PropertyValueFactory<>("priority"));
